third_party/WebKit/LayoutTests/external/wpt/html/dom/self-origin.sub-expected.txt

This is a testharness.js-based test.
PASS Assigning blob url 
PASS Should have the right origin for cross-origin subframe 
PASS Should have the right origin for cross-origin subframe after setting document.domain 
PASS Should have the right origin for IDN subframe 
PASS Should have the right origin for IDN subframe after setting document.domain 
PASS Should have the right origin for sandboxed iframe 
FAIL We should have the right origin for our page assert_equals: expected (string) "http://web-platform.test:8001" but got (undefined) undefined
FAIL about:blank subframe origins assert_equals: Should have the right origin for about:blank iframe expected (string) "http://web-platform.test:8001" but got (undefined) undefined
FAIL blob: subframe origins assert_equals: Should have the right origin for blob: iframe expected (string) "http://web-platform.test:8001" but got (undefined) undefined
FAIL javascript: subframe origins assert_equals: Should have the right origin for javascript: iframe expected (string) "http://web-platform.test:8001" but got (undefined) undefined
FAIL srcdoc subframe origins assert_equals: Should have the right origin for srcdoc iframe expected (string) "http://web-platform.test:8001" but got (undefined) undefined
Harness: the test ran to completion.
